refactor: hsp_sdk_ext の型定義を用意
