+++ src/knowbug_core/hsx_var_metadata.h
@@ -1,0 +1,71 @@
+#pragma once
+
+#include "hsx_types_fwd.h"
+#include "hsx_dim_index.h"
+
+namespace hsp_sdk_ext {
+	// HSP の変数に関連するメタデータの詰め合わせ
+	class HspVarMetadata {
+	public:
+		HspType type_;
+		HspVarMode mode_;
+		HspDimIndex lengths_;
+		std::size_t element_size_;
+		std::size_t data_size_;
+		std::size_t block_size_;
+		void const* data_ptr_;
+		void const* master_ptr_;
+		void const* block_ptr_;
+
+	public:
+		auto type() const -> HspType {
+			return type_;
+		}
+
+		auto mode() const -> HspVarMode {
+			return mode_;
+		}
+
+		auto lengths() const -> HspDimIndex const& {
+			return lengths_;
+		}
+
+		auto element_size() const -> std::size_t {
+			return element_size_;
+		}
+
+		auto data_size() const -> std::size_t {
+			return data_size_;
+		}
+
+		auto block_size() const -> std::size_t {
+			return block_size_;
+		}
+
+		auto data_ptr() const -> void const* {
+			return data_ptr_;
+		}
+
+		auto master_ptr() const -> void const* {
+			return master_ptr_;
+		}
+
+		auto block_ptr() const -> void const* {
+			return block_ptr_;
+		}
+
+		static auto none() -> HspVarMetadata {
+			return HspVarMetadata{
+				HspType::None,
+				HspVarMode::None,
+				HspDimIndex::one(),
+				0,
+				0,
+				0,
+				nullptr,
+				nullptr,
+				nullptr,
+			};
+		}
+	};
+}
