+++ src/knowbug_core/knowbug_core.vcxproj
@@ -351,6 +351,13 @@
     <ClInclude Include="HspTypes.h" />
     <ClInclude Include="hsp_objects_module_tree.h" />
     <ClInclude Include="hsp_wrap_call.h" />
+    <ClInclude Include="hsx_data.h" />
+    <ClInclude Include="hsx_dim_index.h" />
+    <ClInclude Include="hsx.h" />
+    <ClInclude Include="hsx_param_data.h" />
+    <ClInclude Include="hsx_param_stack.h" />
+    <ClInclude Include="hsx_types_fwd.h" />
+    <ClInclude Include="hsx_var_metadata.h" />
     <ClInclude Include="memory_view.h" />
     <ClInclude Include="module\CIni.h" />
     <ClInclude Include="module\CStrWriter.h" />
